requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.1
django-debug-toolbar==4.3.0
Pillow==10.2.0
psycopg-binary==3.1.18
psycopg==3.1.18
python-dotenv==1.0.1
